sqlbase: move DatumAlloc to its own file
Release note: None
